requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic-settings==2.1.0
python-multipart==0.0.6
psutil==5.9.6

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-xdist==3.3.1
pytest-mock==3.12.0
httpx==0.25.2